Added unique prefix to preprocessor macros and updated their usage
Using generic preprocessor macros like HOST opens the risk of another
library or the user trying to define the same macro. To correct this,
I prefixed all macros for the YouTubeAPI with YTAPI_ so they're unique.

Also removed unused macros, replaced the hardcoded hostname with the
existing macro, and pulled out the hardcoded timeout.